Improvements to Model.run() and hyperion wrapper script
It is now possible to run Hyperion in parallel with the mpi= option in Model.run() and the -m option in the hyperion script. When using Model.run(mpi=True), the number of cores can be specified with the n_cores (which defaults to multiprocessing.cpu_count()). Optionally read in MPI command from .hyperionrc config file.