Added methods so as to be able to use this kind of syntax:
b = AMRGrid()
b['density'] = a['density']
b['specific_energy'] = []
b['specific_energy'].append(a['specific_energy'])

This relies on creating 'views' when selecting a specific quantity component, which is basically a 'lite' version of the grid that contains only the selected parameter. Modifications to this view should propagate back to the original grid from which the view was created.