Simplify installation instructions/options
Specifically, this gets rid of many of the individual OS/distribution pages for
installation since it's hard to keep these up to date. Instead, I've opted to
just recommend conda, then provide all the information needed to know which
dependencies to install, without going into detail about how to install them
(this should be explained in the docs for the individual dependencies).

In addition, this gets rid of the cmake compilation configuration since I'm
not able to maintain it and it might go stale/require updating.

Finally, this gets rid of the custom dependency install script since it
requires regular maintenance.